Split `MetaInitializer` test
