+++ tests/tools/map_estimator_test.py
@@ -19,10 +19,10 @@ import pytest
 
 _ = pytest.importorskip("flax")
 
-from ott.geometry import pointcloud  # noqa: E402
-from ott.problems.nn import dataset  # noqa: E402
-from ott.solvers.nn import losses, models  # noqa: E402
-from ott.tools import map_estimator, sinkhorn_divergence  # noqa: E402
+from ott.geometry import pointcloud
+from ott.problems.nn import dataset
+from ott.solvers.nn import losses, models
+from ott.tools import map_estimator, sinkhorn_divergence
 
 
 @pytest.mark.fast()
